requirements.txt
APScheduler==3.9.1
asgiref==3.5.0
async-timeout==4.0.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.3
django-apscheduler==0.6.2
django-cors-headers==3.11.0
django-redis==5.2.0
django-restql==0.15.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.1
MarkupSafe==2.1.1
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
py4j==0.10.9.3
PyJWT==2.3.0
pyparsing==3.0.8
pyPEG2==2.15.2
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.9
user-agents==2.2.0
wrapt==1.14.1
